fixed indexing/table of contents for DB
