+++ database/connection.js
@@ -1,12 +1,18 @@
-const pgp = require('pg-promise')();
+const promise = require('bluebird');
 
-// TBD: Need to set correct config parameters
-const config = {
-  host: 'localhost',
-  port: 3457,
-  database: 'glassCamp',
-  user: 'root',
-  password: ''
+const options = {
+  promiseLib: promise
 };
 
-module.exports.db = pgp(config);
+  // username: 'kennethso',
+  // host: 'localhost',
+  // database: 'campoverde'
+
+
+const pgp = require('pg-promise')(options);
+
+// TBD: Need to set correct config parameters
+
+// const db = pgp('postgres://username:password@host:port/database');
+module.exports.db = pgp('postgres://kennethso:password@localhost/campoverde');
+// console.log("PGP", typeof module.exports.db.query);
